Warning fix for 64-bit systems: u_int64_t is not unsigned long long
git-svn-id: http://madwifi-project.org/svn/madwifi/trunk@3420 0192ed92-7a03-0410-a25b-9323aeb14dbd

+++ tools/ath_info.c
@@ -818,8 +818,8 @@ void dump_timers_register(void *mem, u_int16_t mac_version)
 	printf("TIMER3 : 0x%8x ATIM : %5u TU:%8x(%d)\n", timer3,
 	       timer3 & timer_mask, timer3_tu,
 	       timer3_tu - timer0_tu);
-	printf("TSF    : 0x%8llx TSFTU: %5u TU:%8x\n", now_tsf,
-	       now_tu & timer_mask, now_tu);
+	printf("TSF    : 0x%8llx TSFTU: %5u TU:%8x\n",
+	       (unsigned long long)now_tsf, now_tu & timer_mask, now_tu);
 
 	printf("BEACON: %x\n", AR5K_REG_READ(AR5K_BEACON));
 	printf("LAST_TSTP: %x\n", AR5K_REG_READ(AR5K_LAST_TSTP));
